Debugger/RemoteKd: Add beginnings of debug stub talking the KD remote protocol (WinDbg and friends), heavy work in progress [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@86108 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Debugger/DBGCRemoteKd.cpp
@@ -1682,7 +1682,7 @@ static int dbgcKdCtxPktManipulate64ReadCtrlSpace(PKDCTX pThis, PCKDPACKETMANIPUL
     KDPACKETMANIPULATEHDR RespHdr;
     KDPACKETMANIPULATE_XFERCTRLSPACE64 XferCtrlSpace64;
     uint8_t abResp[sizeof(NTKCONTEXT64)];
-    size_t cbData = 0;
+    uint32_t cbData = 0;
     RT_ZERO(RespHdr); RT_ZERO(XferCtrlSpace64);
     RT_ZERO(abResp);
